+++ Tests/ChessKitTests/GameTests.swift
@@ -91,7 +91,18 @@ class GameTests: XCTestCase {
     }
     
     func testMoveIndexHistory() {
-        let f5History = game.moves.history(for: f5Index)
+        let f5Future = game.moves.history(for: f5Index)
+        
+        XCTAssertEqual(
+            f5Future,
+            [
+                .init(number: 3, color: .white, variation: 3)
+            ]
+        )
+    }
+    
+    func testMoveIndexFuture() {
+        let f5History = game.moves.future(for: f5Index)
         
         XCTAssertEqual(
             f5History,
@@ -99,7 +110,7 @@ class GameTests: XCTestCase {
                 .init(number: 1, color: .white, variation: 0),
                 .init(number: 1, color: .black, variation: 0),
                 .init(number: 2, color: .white, variation: 0),
-                f5Index,
+                f5Index
             ]
         )
     }